tcping_freehostinfo: don't pass NULL to freeaddrinfo()
According to the POSIX man page, "The freeaddrinfo() function shall
free one or more addrinfo structures". Trying to free zero
structures is a misuse of the interface.
tcping segfaults with musl libc if the name does not resolve.
Fix this by checking if `hi->ai` is NULL before calling freeaddrinfo().
Checking whether `hi` is NULL is not strictly necessary, but it makes
the library more robust.